Refactor PageView and SliverFillViewport: Remove shrinkWrapCrossAxis
- Removed the shrinkWrapCrossAxis property from SliverFillViewport and related classes.
- Updated PageView tests to reflect the removal of shrinkWrapCrossAxis, ensuring height adapts to current child.
- Adjusted viewport behavior to maintain expected functionality without shrinkWrapCrossAxis.
- Cleaned up CrossAxisShrinkWrappingViewport class and its associated tests.
