+++ packages/flutter/lib/src/rendering/sliver_fill.dart
@@ -35,13 +35,9 @@ import 'sliver_multi_box_adaptor.dart';
 class RenderSliverFillViewport extends RenderSliverFixedExtentBoxAdaptor {
   /// Creates a sliver that contains multiple box children that each fill the
   /// viewport.
-  RenderSliverFillViewport({
-    required super.childManager,
-    double viewportFraction = 1.0,
-    bool shrinkWrapCrossAxis = false,
-  }) : assert(viewportFraction > 0.0),
-       _viewportFraction = viewportFraction,
-       _shrinkWrapCrossAxis = shrinkWrapCrossAxis;
+  RenderSliverFillViewport({required super.childManager, double viewportFraction = 1.0})
+    : assert(viewportFraction > 0.0),
+      _viewportFraction = viewportFraction;
 
   @override
   double get itemExtent => constraints.viewportMainAxisExtent * viewportFraction;
@@ -59,258 +55,6 @@ class RenderSliverFillViewport extends RenderSliverFixedExtentBoxAdaptor {
     }
     _viewportFraction = value;
     markNeedsLayout();
-  }
-
-  /// Whether the sliver should allow each child to determine its own cross-axis
-  /// extent and report the visible page's extent in [geometry].
-  ///
-  /// When false, every child is forced to fill the viewport in the cross axis.
-  bool get shrinkWrapCrossAxis => _shrinkWrapCrossAxis;
-  bool _shrinkWrapCrossAxis;
-  set shrinkWrapCrossAxis(bool value) {
-    if (_shrinkWrapCrossAxis == value) {
-      return;
-    }
-    _shrinkWrapCrossAxis = value;
-    markNeedsLayout();
-  }
-
-  @override
-  void performLayout() {
-    if (!shrinkWrapCrossAxis) {
-      super.performLayout();
-      return;
-    }
-
-    final SliverConstraints constraints = this.constraints;
-    childManager.didStartLayout();
-    childManager.setDidUnderflow(false);
-
-    final double scrollOffset = constraints.scrollOffset + constraints.cacheOrigin;
-    assert(scrollOffset >= 0.0);
-    final double remainingExtent = constraints.remainingCacheExtent;
-    assert(remainingExtent >= 0.0);
-    final double targetEndScrollOffset = scrollOffset + remainingExtent;
-
-    // TODO(Piinks): Clean up when deprecation expires.
-    const double deprecatedExtraItemExtent = -1;
-
-    final int firstIndex = getMinChildIndexForScrollOffset(scrollOffset, deprecatedExtraItemExtent);
-    final int? targetLastIndex = targetEndScrollOffset.isFinite
-        ? getMaxChildIndexForScrollOffset(targetEndScrollOffset, deprecatedExtraItemExtent)
-        : null;
-
-    if (firstChild != null) {
-      final int leadingGarbage = calculateLeadingGarbage(firstIndex: firstIndex);
-      final int trailingGarbage = targetLastIndex != null
-          ? calculateTrailingGarbage(lastIndex: targetLastIndex)
-          : 0;
-      collectGarbage(leadingGarbage, trailingGarbage);
-    } else {
-      collectGarbage(0, 0);
-    }
-
-    if (firstChild == null) {
-      final double layoutOffset = indexToLayoutOffset(deprecatedExtraItemExtent, firstIndex);
-      if (!addInitialChild(index: firstIndex, layoutOffset: layoutOffset)) {
-        final double max = firstIndex <= 0
-            ? 0.0
-            : computeMaxScrollOffset(constraints, deprecatedExtraItemExtent);
-        geometry = SliverGeometry(scrollExtent: max, maxPaintExtent: max, crossAxisExtent: 0.0);
-        childManager.didFinishLayout();
-        return;
-      }
-    }
-
-    final BoxConstraints childConstraints = _getShrinkWrapCrossAxisChildConstraints();
-
-    RenderBox? trailingChildWithLayout;
-
-    for (int index = indexOf(firstChild!) - 1; index >= firstIndex; --index) {
-      final RenderBox? child = insertAndLayoutLeadingChild(childConstraints, parentUsesSize: true);
-      if (child == null) {
-        geometry = SliverGeometry(
-          scrollOffsetCorrection: indexToLayoutOffset(deprecatedExtraItemExtent, index),
-        );
-        return;
-      }
-      final SliverMultiBoxAdaptorParentData childParentData =
-          child.parentData! as SliverMultiBoxAdaptorParentData;
-      childParentData.layoutOffset = indexToLayoutOffset(deprecatedExtraItemExtent, index);
-      assert(childParentData.index == index);
-      trailingChildWithLayout ??= child;
-    }
-
-    if (trailingChildWithLayout == null) {
-      firstChild!.layout(childConstraints, parentUsesSize: true);
-      final SliverMultiBoxAdaptorParentData childParentData =
-          firstChild!.parentData! as SliverMultiBoxAdaptorParentData;
-      childParentData.layoutOffset = indexToLayoutOffset(deprecatedExtraItemExtent, firstIndex);
-      trailingChildWithLayout = firstChild;
-    }
-
-    double estimatedMaxScrollOffset = double.infinity;
-    for (
-      int index = indexOf(trailingChildWithLayout!) + 1;
-      targetLastIndex == null || index <= targetLastIndex;
-      ++index
-    ) {
-      RenderBox? child = childAfter(trailingChildWithLayout!);
-      if (child == null || indexOf(child) != index) {
-        child = insertAndLayoutChild(
-          childConstraints,
-          after: trailingChildWithLayout,
-          parentUsesSize: true,
-        );
-        if (child == null) {
-          estimatedMaxScrollOffset = indexToLayoutOffset(deprecatedExtraItemExtent, index);
-          break;
-        }
-      } else {
-        child.layout(childConstraints, parentUsesSize: true);
-      }
-      trailingChildWithLayout = child;
-      final SliverMultiBoxAdaptorParentData childParentData =
-          child.parentData! as SliverMultiBoxAdaptorParentData;
-      assert(childParentData.index == index);
-      childParentData.layoutOffset = indexToLayoutOffset(
-        deprecatedExtraItemExtent,
-        childParentData.index!,
-      );
-    }
-
-    final int lastIndex = indexOf(lastChild!);
-    final double leadingScrollOffset = indexToLayoutOffset(deprecatedExtraItemExtent, firstIndex);
-    final double trailingScrollOffset = indexToLayoutOffset(
-      deprecatedExtraItemExtent,
-      lastIndex + 1,
-    );
-
-    assert(
-      firstIndex == 0 || childScrollOffset(firstChild!)! - scrollOffset <= precisionErrorTolerance,
-    );
-    assert(debugAssertChildListIsNonEmptyAndContiguous());
-    assert(indexOf(firstChild!) == firstIndex);
-    assert(targetLastIndex == null || lastIndex <= targetLastIndex);
-
-    estimatedMaxScrollOffset = math.min(
-      estimatedMaxScrollOffset,
-      estimateMaxScrollOffset(
-        constraints,
-        firstIndex: firstIndex,
-        lastIndex: lastIndex,
-        leadingScrollOffset: leadingScrollOffset,
-        trailingScrollOffset: trailingScrollOffset,
-      ),
-    );
-
-    final double paintExtent = calculatePaintOffset(
-      constraints,
-      from: leadingScrollOffset,
-      to: trailingScrollOffset,
-    );
-
-    final double cacheExtent = calculateCacheOffset(
-      constraints,
-      from: leadingScrollOffset,
-      to: trailingScrollOffset,
-    );
-
-    final double targetEndScrollOffsetForPaint =
-        constraints.scrollOffset + constraints.remainingPaintExtent;
-    final int? targetLastIndexForPaint = targetEndScrollOffsetForPaint.isFinite
-        ? getMaxChildIndexForScrollOffset(targetEndScrollOffsetForPaint, deprecatedExtraItemExtent)
-        : null;
-
-    geometry = SliverGeometry(
-      scrollExtent: estimatedMaxScrollOffset,
-      paintExtent: paintExtent,
-      cacheExtent: cacheExtent,
-      maxPaintExtent: estimatedMaxScrollOffset,
-      crossAxisExtent: _currentPageCrossAxisExtent(),
-      hasVisualOverflow:
-          (targetLastIndexForPaint != null && lastIndex >= targetLastIndexForPaint) ||
-          constraints.scrollOffset > 0.0,
-    );
-
-    if (estimatedMaxScrollOffset == trailingScrollOffset) {
-      childManager.setDidUnderflow(true);
-    }
-    childManager.didFinishLayout();
-  }
-
-  BoxConstraints _getShrinkWrapCrossAxisChildConstraints() {
-    final double extent = itemExtent;
-    return switch (constraints.axis) {
-      Axis.horizontal => BoxConstraints(
-        minWidth: extent,
-        maxWidth: extent,
-        minHeight: 0.0,
-        maxHeight: constraints.crossAxisExtent,
-      ),
-      Axis.vertical => BoxConstraints(
-        minWidth: 0.0,
-        maxWidth: constraints.crossAxisExtent,
-        minHeight: extent,
-        maxHeight: extent,
-      ),
-    };
-  }
-
-  RenderBox? _childForIndex(int index) {
-    RenderBox? child = firstChild;
-    while (child != null) {
-      final int childIndex = indexOf(child);
-      if (childIndex == index) {
-        return child;
-      }
-      if (childIndex > index) {
-        return null;
-      }
-      child = childAfter(child);
-    }
-    return null;
-  }
-
-  double _childCrossAxisExtent(RenderBox child) {
-    return switch (constraints.axis) {
-      Axis.horizontal => child.size.height,
-      Axis.vertical => child.size.width,
-    };
-  }
-
-  double _currentPageCrossAxisExtent() {
-    final RenderBox? firstLiveChild = firstChild;
-    if (firstLiveChild == null) {
-      return 0.0;
-    }
-
-    final double mainAxisExtent = itemExtent;
-    if (mainAxisExtent == 0.0) {
-      return _childCrossAxisExtent(firstLiveChild);
-    }
-
-    final double initialPageOffset = math.max(
-      0.0,
-      (mainAxisExtent - constraints.viewportMainAxisExtent) / 2.0,
-    );
-    final double rawPage =
-        math.max(0.0, constraints.scrollOffset - initialPageOffset) / mainAxisExtent;
-    final int lowerIndex = rawPage.floor();
-    final int upperIndex = rawPage.ceil();
-    final RenderBox? lowerChild = _childForIndex(lowerIndex);
-    final RenderBox? upperChild = _childForIndex(upperIndex);
-
-    if (lowerChild == null && upperChild == null) {
-      return _childCrossAxisExtent(firstLiveChild);
-    }
-
-    final double lowerExtent = lowerChild != null
-        ? _childCrossAxisExtent(lowerChild)
-        : _childCrossAxisExtent(upperChild!);
-    final double upperExtent = upperChild != null ? _childCrossAxisExtent(upperChild) : lowerExtent;
-    final double pageDelta = rawPage - lowerIndex;
-    return lowerExtent + (upperExtent - lowerExtent) * pageDelta;
   }
 }
