+++ packages/flutter/lib/src/rendering/viewport.dart
@@ -2243,285 +2243,3 @@ class RenderShrinkWrappingViewport extends RenderViewportBase<SliverLogicalConta
   @override
   String labelForChild(int index) => 'child $index';
 }
-
-/// A render object that is bigger on the inside and shrink-wraps its children
-/// in the cross axis.
-///
-/// [RenderCrossAxisShrinkWrappingViewport] displays a subset of its children
-/// according to its own dimensions and the given [offset]. As the offset
-/// varies, different children are visible through the viewport.
-///
-/// [RenderCrossAxisShrinkWrappingViewport] differs from [RenderViewport] in
-/// that [RenderViewport] expands to fill the cross axis whereas
-/// [RenderCrossAxisShrinkWrappingViewport] sizes itself to match its children
-/// in the cross axis. This shrink wrapping behavior is expensive because the
-/// children, and hence the viewport, could potentially change size whenever the
-/// [offset] changes.
-///
-/// [RenderCrossAxisShrinkWrappingViewport] cannot contain [RenderBox] children
-/// directly. Instead, use a [RenderSliverList], [RenderSliverFixedExtentList],
-/// [RenderSliverGrid], or a [RenderSliverToBoxAdapter], for example.
-///
-/// See also:
-///
-///  * [RenderViewport], a viewport that does not shrink-wrap its contents.
-///  * [RenderShrinkWrappingViewport], a viewport that shrink-wraps in the main
-///    axis.
-///  * [RenderSliver], which explains more about the Sliver protocol.
-///  * [RenderBox], which explains more about the Box protocol.
-///  * [RenderSliverToBoxAdapter], which allows a [RenderBox] object to be
-///    placed inside a [RenderSliver] (the opposite of this class).
-class RenderCrossAxisShrinkWrappingViewport
-    extends RenderViewportBase<SliverLogicalContainerParentData> {
-  /// Creates a viewport (for [RenderSliver] objects) that shrink-wraps its
-  /// contents in the cross axis.
-  ///
-  /// The [offset] must be specified. For testing purposes, consider passing a
-  /// [ViewportOffset.zero] or [ViewportOffset.fixed].
-  RenderCrossAxisShrinkWrappingViewport({
-    super.axisDirection,
-    required super.crossAxisDirection,
-    required super.offset,
-    super.paintOrder,
-    super.clipBehavior,
-    super.scrollCacheExtent,
-    List<RenderSliver>? children,
-  }) {
-    addAll(children);
-  }
-
-  @override
-  void setupParentData(RenderObject child) {
-    if (child.parentData is! SliverLogicalContainerParentData) {
-      child.parentData = SliverLogicalContainerParentData();
-    }
-  }
-
-  @override
-  bool debugThrowIfNotCheckingIntrinsics() {
-    assert(() {
-      if (!RenderObject.debugCheckingIntrinsics) {
-        throw FlutterError.fromParts(<DiagnosticsNode>[
-          ErrorSummary('$runtimeType does not support returning intrinsic dimensions.'),
-          ErrorDescription(
-            'Calculating the intrinsic dimensions would require instantiating every child of '
-            'the viewport, which defeats the point of viewports being lazy.',
-          ),
-          ErrorHint(
-            'If you are merely trying to shrink-wrap the viewport in the cross axis direction, '
-            'you should be able to achieve that effect by just giving the viewport loose '
-            'constraints, without needing to measure its intrinsic dimensions.',
-          ),
-        ]);
-      }
-      return true;
-    }());
-    return true;
-  }
-
-  late double _maxScrollExtent;
-  late double _shrinkWrapExtent;
-  bool _hasVisualOverflow = false;
-
-  bool _debugCheckHasBoundedMainAxis() {
-    assert(() {
-      switch (axis) {
-        case Axis.vertical:
-          if (!constraints.hasBoundedHeight) {
-            throw FlutterError(
-              'Vertical viewport was given unbounded height.\n'
-              'Cross-axis-shrinkwrapping viewports expand in the main axis to fill their '
-              'container and constrain their children to match their extent in the main axis. '
-              'In this case, a vertical cross-axis-shrinkwrapping viewport was given an '
-              'unlimited amount of vertical space in which to expand.',
-            );
-          }
-        case Axis.horizontal:
-          if (!constraints.hasBoundedWidth) {
-            throw FlutterError(
-              'Horizontal viewport was given unbounded width.\n'
-              'Cross-axis-shrinkwrapping viewports expand in the main axis to fill their '
-              'container and constrain their children to match their extent in the main axis. '
-              'In this case, a horizontal cross-axis-shrinkwrapping viewport was given an '
-              'unlimited amount of horizontal space in which to expand.',
-            );
-          }
-      }
-      return true;
-    }());
-    return true;
-  }
-
-  @override
-  void performLayout() {
-    final BoxConstraints constraints = this.constraints;
-    assert(_debugCheckHasBoundedMainAxis());
-
-    final (double mainAxisExtent, double crossAxisExtent) = switch (axis) {
-      Axis.vertical => (constraints.maxHeight, constraints.maxWidth),
-      Axis.horizontal => (constraints.maxWidth, constraints.maxHeight),
-    };
-
-    offset.applyViewportDimension(mainAxisExtent);
-
-    if (firstChild == null) {
-      size = switch (axis) {
-        Axis.vertical => Size(constraints.minWidth, mainAxisExtent),
-        Axis.horizontal => Size(mainAxisExtent, constraints.minHeight),
-      };
-      _maxScrollExtent = 0.0;
-      _shrinkWrapExtent = 0.0;
-      _hasVisualOverflow = false;
-      offset.applyContentDimensions(0.0, 0.0);
-      return;
-    }
-
-    double correction;
-    while (true) {
-      correction = _attemptLayout(mainAxisExtent, crossAxisExtent, offset.pixels);
-      if (correction != 0.0) {
-        offset.correctBy(correction);
-      } else {
-        final bool didAcceptContentDimension = offset.applyContentDimensions(
-          0.0,
-          math.max(0.0, _maxScrollExtent - mainAxisExtent),
-        );
-        if (didAcceptContentDimension) {
-          break;
-        }
-      }
-    }
-
-    final double effectiveExtent = switch (axis) {
-      Axis.vertical => constraints.constrainWidth(_shrinkWrapExtent),
-      Axis.horizontal => constraints.constrainHeight(_shrinkWrapExtent),
-    };
-
-    size = switch (axis) {
-      Axis.vertical => constraints.constrainDimensions(effectiveExtent, mainAxisExtent),
-      Axis.horizontal => constraints.constrainDimensions(mainAxisExtent, effectiveExtent),
-    };
-  }
-
-  double _attemptLayout(double mainAxisExtent, double crossAxisExtent, double correctedOffset) {
-    assert(!mainAxisExtent.isNaN);
-    assert(mainAxisExtent >= 0.0);
-    assert(mainAxisExtent.isFinite);
-    assert(!crossAxisExtent.isNaN);
-    assert(crossAxisExtent >= 0.0);
-    assert(correctedOffset.isFinite);
-    _maxScrollExtent = 0.0;
-    _shrinkWrapExtent = 0.0;
-    _hasVisualOverflow = correctedOffset < 0.0;
-
-    _calculatedCacheExtent = _scrollCacheExtent._calculateCacheOffset(mainAxisExtent);
-
-    return layoutChildSequence(
-      child: firstChild,
-      scrollOffset: math.max(0.0, correctedOffset),
-      overlap: math.min(0.0, correctedOffset),
-      layoutOffset: math.max(0.0, -correctedOffset),
-      remainingPaintExtent: mainAxisExtent + math.min(0.0, correctedOffset),
-      mainAxisExtent: mainAxisExtent,
-      crossAxisExtent: crossAxisExtent,
-      growthDirection: GrowthDirection.forward,
-      advance: childAfter,
-      remainingCacheExtent: mainAxisExtent + 2 * _calculatedCacheExtent!,
-      cacheOrigin: -_calculatedCacheExtent!,
-    );
-  }
-
-  @override
-  bool get hasVisualOverflow => _hasVisualOverflow;
-
-  @override
-  void updateOutOfBandData(GrowthDirection growthDirection, SliverGeometry childLayoutGeometry) {
-    assert(growthDirection == GrowthDirection.forward);
-    _maxScrollExtent += childLayoutGeometry.scrollExtent;
-    if (childLayoutGeometry.hasVisualOverflow) {
-      _hasVisualOverflow = true;
-    }
-    final double childCrossAxisExtent =
-        childLayoutGeometry.crossAxisExtent ??
-        switch (axis) {
-          Axis.vertical => constraints.maxWidth,
-          Axis.horizontal => constraints.maxHeight,
-        };
-    assert(childCrossAxisExtent.isFinite);
-    _shrinkWrapExtent = math.max(_shrinkWrapExtent, childCrossAxisExtent);
-  }
-
-  @override
-  void updateChildLayoutOffset(
-    RenderSliver child,
-    double layoutOffset,
-    GrowthDirection growthDirection,
-  ) {
-    assert(growthDirection == GrowthDirection.forward);
-    final SliverLogicalParentData childParentData = child.parentData! as SliverLogicalParentData;
-    childParentData.layoutOffset = layoutOffset;
-  }
-
-  @override
-  Offset paintOffsetOf(RenderSliver child) {
-    final SliverLogicalParentData childParentData = child.parentData! as SliverLogicalParentData;
-    return computeAbsolutePaintOffset(
-      child,
-      childParentData.layoutOffset!,
-      GrowthDirection.forward,
-    );
-  }
-
-  @override
-  double scrollOffsetOf(RenderSliver child, double scrollOffsetWithinChild) {
-    assert(child.parent == this);
-    assert(child.constraints.growthDirection == GrowthDirection.forward);
-    var scrollOffsetToChild = 0.0;
-    RenderSliver? current = firstChild;
-    while (current != child) {
-      scrollOffsetToChild += current!.geometry!.scrollExtent;
-      current = childAfter(current);
-    }
-    return scrollOffsetToChild + scrollOffsetWithinChild;
-  }
-
-  @override
-  double maxScrollObstructionExtentBefore(RenderSliver child) {
-    assert(child.parent == this);
-    assert(child.constraints.growthDirection == GrowthDirection.forward);
-    var pinnedExtent = 0.0;
-    RenderSliver? current = firstChild;
-    while (current != child) {
-      pinnedExtent += current!.geometry!.maxScrollObstructionExtent;
-      current = childAfter(current);
-    }
-    return pinnedExtent;
-  }
-
-  @override
-  void applyPaintTransform(RenderObject child, Matrix4 transform) {
-    // Hit test logic relies on this always providing an invertible matrix.
-    final Offset offset = paintOffsetOf(child as RenderSliver);
-    transform.translateByDouble(offset.dx, offset.dy, 0, 1);
-  }
-
-  @override
-  double computeChildMainAxisPosition(RenderSliver child, double parentMainAxisPosition) {
-    assert(hasSize);
-    final double layoutOffset = (child.parentData! as SliverLogicalParentData).layoutOffset!;
-    return switch (applyGrowthDirectionToAxisDirection(
-      child.constraints.axisDirection,
-      child.constraints.growthDirection,
-    )) {
-      AxisDirection.down || AxisDirection.right => parentMainAxisPosition - layoutOffset,
-      AxisDirection.up => size.height - parentMainAxisPosition - layoutOffset,
-      AxisDirection.left => size.width - parentMainAxisPosition - layoutOffset,
-    };
-  }
-
-  @override
-  int get indexOfFirstChild => 0;
-
-  @override
-  String labelForChild(int index) => 'child $index';
-}
