+++ packages/flutter/lib/src/widgets/page_view.dart
@@ -31,6 +31,8 @@ import 'scroll_position_with_single_context.dart';
 import 'scroll_view.dart';
 import 'scrollable.dart';
 import 'sliver_fill.dart';
+import 'sliver.dart';
+import 'sliver_layout_builder.dart';
 import 'viewport.dart';
 
 /// A controller for [PageView].
@@ -900,6 +902,10 @@ class PageView extends StatefulWidget {
   /// behavior because the viewport can change size whenever the page position
   /// changes.
   ///
+  /// Like any [PageView], the scrolling axis must still be bounded by the
+  /// parent. For a horizontal [PageView], this means the width must be bounded.
+  /// For a vertical [PageView], the height must be bounded.
+  ///
   /// {@tool dartpad}
   /// This example shows a [PageView] with [shrinkWrapCrossAxis] set to true,
   /// placed inside a [Column]. Each page has a different height, and the
@@ -999,7 +1005,7 @@ class _PageViewState extends State<PageView> {
             widget.scrollBehavior ?? ScrollConfiguration.of(context).copyWith(scrollbars: false),
         viewportBuilder: (BuildContext context, ViewportOffset position) {
           final Widget viewport = widget.shrinkWrapCrossAxis
-              ? CrossAxisShrinkWrappingViewport(
+              ? _PageViewCrossAxisShrinkWrappingViewport(
                   scrollCacheExtent: widget.allowImplicitScrolling
                       ? const ScrollCacheExtent.viewport(1.0)
                       : const ScrollCacheExtent.viewport(0.0),
@@ -1007,11 +1013,10 @@ class _PageViewState extends State<PageView> {
                   offset: position,
                   clipBehavior: widget.clipBehavior,
                   slivers: <Widget>[
-                    SliverFillViewport(
+                    _PageViewShrinkWrappingSliverFillViewport(
                       viewportFraction: _controller.viewportFraction,
                       delegate: widget.childrenDelegate,
                       padEnds: widget.padEnds,
-                      shrinkWrapCrossAxis: true,
                     ),
                   ],
                 )
@@ -1065,5 +1070,592 @@ class _PageViewState extends State<PageView> {
         ifTrue: 'cross-axis shrink-wrapping',
       ),
     );
+  }
+}
+
+class _PageViewShrinkWrappingSliverFillViewport extends StatelessWidget {
+  const _PageViewShrinkWrappingSliverFillViewport({
+    required this.delegate,
+    required this.viewportFraction,
+    required this.padEnds,
+  });
+
+  final SliverChildDelegate delegate;
+  final double viewportFraction;
+  final bool padEnds;
+
+  @override
+  Widget build(BuildContext context) {
+    final double paddingFraction = padEnds ? clampDouble(1 - viewportFraction, 0, 1) / 2 : 0;
+    final Widget sliver = _PageViewSliverFillViewportRenderObjectWidget(
+      viewportFraction: viewportFraction,
+      delegate: delegate,
+    );
+    if (paddingFraction == 0.0) {
+      return sliver;
+    }
+    return SliverLayoutBuilder(
+      builder: (BuildContext context, SliverConstraints constraints) {
+        final double paddingValue = constraints.viewportMainAxisExtent * paddingFraction;
+        return SliverPadding(
+          padding: switch (constraints.axis) {
+            Axis.horizontal => EdgeInsets.symmetric(horizontal: paddingValue),
+            Axis.vertical => EdgeInsets.symmetric(vertical: paddingValue),
+          },
+          sliver: sliver,
+        );
+      },
+    );
+  }
+}
+
+class _PageViewSliverFillViewportRenderObjectWidget extends SliverMultiBoxAdaptorWidget {
+  const _PageViewSliverFillViewportRenderObjectWidget({
+    required super.delegate,
+    this.viewportFraction = 1.0,
+  }) : assert(viewportFraction > 0.0);
+
+  final double viewportFraction;
+
+  @override
+  _RenderSliverFillViewportWithCrossAxisShrinkWrapping createRenderObject(BuildContext context) {
+    final SliverMultiBoxAdaptorElement element = context as SliverMultiBoxAdaptorElement;
+    return _RenderSliverFillViewportWithCrossAxisShrinkWrapping(
+      childManager: element,
+      viewportFraction: viewportFraction,
+    );
+  }
+
+  @override
+  void updateRenderObject(
+    BuildContext context,
+    _RenderSliverFillViewportWithCrossAxisShrinkWrapping renderObject,
+  ) {
+    renderObject.viewportFraction = viewportFraction;
+  }
+}
+
+class _RenderSliverFillViewportWithCrossAxisShrinkWrapping extends RenderSliverFillViewport {
+  _RenderSliverFillViewportWithCrossAxisShrinkWrapping({
+    required super.childManager,
+    double viewportFraction = 1.0,
+  }) : super(viewportFraction: viewportFraction);
+
+  @override
+  void performLayout() {
+    final SliverConstraints constraints = this.constraints;
+    childManager.didStartLayout();
+    childManager.setDidUnderflow(false);
+
+    final double scrollOffset = constraints.scrollOffset + constraints.cacheOrigin;
+    assert(scrollOffset >= 0.0);
+    final double remainingExtent = constraints.remainingCacheExtent;
+    assert(remainingExtent >= 0.0);
+    final double targetEndScrollOffset = scrollOffset + remainingExtent;
+
+    // TODO(Piinks): Clean up when deprecation expires.
+    const double deprecatedExtraItemExtent = -1;
+
+    final int firstIndex = getMinChildIndexForScrollOffset(scrollOffset, deprecatedExtraItemExtent);
+    final int? targetLastIndex = targetEndScrollOffset.isFinite
+        ? getMaxChildIndexForScrollOffset(targetEndScrollOffset, deprecatedExtraItemExtent)
+        : null;
+
+    if (firstChild != null) {
+      final int leadingGarbage = calculateLeadingGarbage(firstIndex: firstIndex);
+      final int trailingGarbage = targetLastIndex != null
+          ? calculateTrailingGarbage(lastIndex: targetLastIndex)
+          : 0;
+      collectGarbage(leadingGarbage, trailingGarbage);
+    } else {
+      collectGarbage(0, 0);
+    }
+
+    if (firstChild == null) {
+      final double layoutOffset = indexToLayoutOffset(deprecatedExtraItemExtent, firstIndex);
+      if (!addInitialChild(index: firstIndex, layoutOffset: layoutOffset)) {
+        final double max = firstIndex <= 0
+            ? 0.0
+            : computeMaxScrollOffset(constraints, deprecatedExtraItemExtent);
+        geometry = SliverGeometry(scrollExtent: max, maxPaintExtent: max, crossAxisExtent: 0.0);
+        childManager.didFinishLayout();
+        return;
+      }
+    }
+
+    final BoxConstraints childConstraints = _getChildConstraintsForShrinkWrapping();
+
+    RenderBox? trailingChildWithLayout;
+
+    for (int index = indexOf(firstChild!) - 1; index >= firstIndex; --index) {
+      final RenderBox? child = insertAndLayoutLeadingChild(childConstraints, parentUsesSize: true);
+      if (child == null) {
+        geometry = SliverGeometry(
+          scrollOffsetCorrection: indexToLayoutOffset(deprecatedExtraItemExtent, index),
+        );
+        return;
+      }
+      final SliverMultiBoxAdaptorParentData childParentData =
+          child.parentData! as SliverMultiBoxAdaptorParentData;
+      childParentData.layoutOffset = indexToLayoutOffset(deprecatedExtraItemExtent, index);
+      assert(childParentData.index == index);
+      trailingChildWithLayout ??= child;
+    }
+
+    if (trailingChildWithLayout == null) {
+      firstChild!.layout(childConstraints, parentUsesSize: true);
+      final SliverMultiBoxAdaptorParentData childParentData =
+          firstChild!.parentData! as SliverMultiBoxAdaptorParentData;
+      childParentData.layoutOffset = indexToLayoutOffset(deprecatedExtraItemExtent, firstIndex);
+      trailingChildWithLayout = firstChild;
+    }
+
+    double estimatedMaxScrollOffset = double.infinity;
+    for (
+      int index = indexOf(trailingChildWithLayout!) + 1;
+      targetLastIndex == null || index <= targetLastIndex;
+      ++index
+    ) {
+      RenderBox? child = childAfter(trailingChildWithLayout!);
+      if (child == null || indexOf(child) != index) {
+        child = insertAndLayoutChild(
+          childConstraints,
+          after: trailingChildWithLayout,
+          parentUsesSize: true,
+        );
+        if (child == null) {
+          estimatedMaxScrollOffset = indexToLayoutOffset(deprecatedExtraItemExtent, index);
+          break;
+        }
+      } else {
+        child.layout(childConstraints, parentUsesSize: true);
+      }
+      trailingChildWithLayout = child;
+      final SliverMultiBoxAdaptorParentData childParentData =
+          child.parentData! as SliverMultiBoxAdaptorParentData;
+      assert(childParentData.index == index);
+      childParentData.layoutOffset = indexToLayoutOffset(
+        deprecatedExtraItemExtent,
+        childParentData.index!,
+      );
+    }
+
+    final int lastIndex = indexOf(lastChild!);
+    final double leadingScrollOffset = indexToLayoutOffset(deprecatedExtraItemExtent, firstIndex);
+    final double trailingScrollOffset = indexToLayoutOffset(
+      deprecatedExtraItemExtent,
+      lastIndex + 1,
+    );
+
+    assert(
+      firstIndex == 0 || childScrollOffset(firstChild!)! - scrollOffset <= precisionErrorTolerance,
+    );
+    assert(debugAssertChildListIsNonEmptyAndContiguous());
+    assert(indexOf(firstChild!) == firstIndex);
+    assert(targetLastIndex == null || lastIndex <= targetLastIndex);
+
+    estimatedMaxScrollOffset = math.min(
+      estimatedMaxScrollOffset,
+      estimateMaxScrollOffset(
+        constraints,
+        firstIndex: firstIndex,
+        lastIndex: lastIndex,
+        leadingScrollOffset: leadingScrollOffset,
+        trailingScrollOffset: trailingScrollOffset,
+      ),
+    );
+
+    final double paintExtent = calculatePaintOffset(
+      constraints,
+      from: leadingScrollOffset,
+      to: trailingScrollOffset,
+    );
+
+    final double cacheExtent = calculateCacheOffset(
+      constraints,
+      from: leadingScrollOffset,
+      to: trailingScrollOffset,
+    );
+
+    final double targetEndScrollOffsetForPaint =
+        constraints.scrollOffset + constraints.remainingPaintExtent;
+    final int? targetLastIndexForPaint = targetEndScrollOffsetForPaint.isFinite
+        ? getMaxChildIndexForScrollOffset(targetEndScrollOffsetForPaint, deprecatedExtraItemExtent)
+        : null;
+
+    geometry = SliverGeometry(
+      scrollExtent: estimatedMaxScrollOffset,
+      paintExtent: paintExtent,
+      cacheExtent: cacheExtent,
+      maxPaintExtent: estimatedMaxScrollOffset,
+      crossAxisExtent: _currentPageCrossAxisExtent(),
+      hasVisualOverflow:
+          (targetLastIndexForPaint != null && lastIndex >= targetLastIndexForPaint) ||
+          constraints.scrollOffset > 0.0,
+    );
+
+    if (estimatedMaxScrollOffset == trailingScrollOffset) {
+      childManager.setDidUnderflow(true);
+    }
+    childManager.didFinishLayout();
+  }
+
+  BoxConstraints _getChildConstraintsForShrinkWrapping() {
+    final double extent = itemExtent;
+    return switch (constraints.axis) {
+      Axis.horizontal => BoxConstraints(
+        minWidth: extent,
+        maxWidth: extent,
+        minHeight: 0.0,
+        maxHeight: constraints.crossAxisExtent,
+      ),
+      Axis.vertical => BoxConstraints(
+        minWidth: 0.0,
+        maxWidth: constraints.crossAxisExtent,
+        minHeight: extent,
+        maxHeight: extent,
+      ),
+    };
+  }
+
+  RenderBox? _childForIndex(int index) {
+    RenderBox? child = firstChild;
+    while (child != null) {
+      final int childIndex = indexOf(child);
+      if (childIndex == index) {
+        return child;
+      }
+      if (childIndex > index) {
+        return null;
+      }
+      child = childAfter(child);
+    }
+    return null;
+  }
+
+  double _childCrossAxisExtent(RenderBox child) {
+    return switch (constraints.axis) {
+      Axis.horizontal => child.size.height,
+      Axis.vertical => child.size.width,
+    };
+  }
+
+  double _currentPageCrossAxisExtent() {
+    final RenderBox? firstLiveChild = firstChild;
+    if (firstLiveChild == null) {
+      return 0.0;
+    }
+
+    final double mainAxisExtent = itemExtent;
+    if (mainAxisExtent == 0.0) {
+      return _childCrossAxisExtent(firstLiveChild);
+    }
+
+    final double initialPageOffset = math.max(
+      0.0,
+      (mainAxisExtent - constraints.viewportMainAxisExtent) / 2.0,
+    );
+    final double rawPage =
+        math.max(0.0, constraints.scrollOffset - initialPageOffset) / mainAxisExtent;
+    final int lowerIndex = rawPage.floor();
+    final int upperIndex = rawPage.ceil();
+    final RenderBox? lowerChild = _childForIndex(lowerIndex);
+    final RenderBox? upperChild = _childForIndex(upperIndex);
+
+    if (lowerChild == null && upperChild == null) {
+      return _childCrossAxisExtent(firstLiveChild);
+    }
+
+    final double lowerExtent = lowerChild != null
+        ? _childCrossAxisExtent(lowerChild)
+        : _childCrossAxisExtent(upperChild!);
+    final double upperExtent = upperChild != null ? _childCrossAxisExtent(upperChild) : lowerExtent;
+    final double pageDelta = rawPage - lowerIndex;
+    return lowerExtent + (upperExtent - lowerExtent) * pageDelta;
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Private cross-axis-shrink-wrapping viewport used only by PageView.
+// ---------------------------------------------------------------------------
+
+class _PageViewCrossAxisShrinkWrappingViewport extends MultiChildRenderObjectWidget {
+  const _PageViewCrossAxisShrinkWrappingViewport({
+    this.axisDirection = AxisDirection.down,
+    required this.offset,
+    this.clipBehavior = Clip.hardEdge,
+    this.scrollCacheExtent,
+    List<Widget> slivers = const <Widget>[],
+  }) : super(children: slivers);
+
+  final AxisDirection axisDirection;
+  final ViewportOffset offset;
+  final Clip clipBehavior;
+  final ScrollCacheExtent? scrollCacheExtent;
+
+  @override
+  MultiChildRenderObjectElement createElement() =>
+      _PageViewCrossAxisShrinkWrappingViewportElement(this);
+
+  @override
+  _RenderPageViewCrossAxisShrinkWrappingViewport createRenderObject(BuildContext context) {
+    return _RenderPageViewCrossAxisShrinkWrappingViewport(
+      axisDirection: axisDirection,
+      crossAxisDirection: Viewport.getDefaultCrossAxisDirection(context, axisDirection),
+      offset: offset,
+      clipBehavior: clipBehavior,
+      scrollCacheExtent: scrollCacheExtent,
+    );
+  }
+
+  @override
+  void updateRenderObject(
+    BuildContext context,
+    _RenderPageViewCrossAxisShrinkWrappingViewport renderObject,
+  ) {
+    renderObject
+      ..axisDirection = axisDirection
+      ..crossAxisDirection = Viewport.getDefaultCrossAxisDirection(context, axisDirection)
+      ..offset = offset
+      ..clipBehavior = clipBehavior
+      ..scrollCacheExtent = scrollCacheExtent;
+  }
+}
+
+class _PageViewCrossAxisShrinkWrappingViewportElement extends MultiChildRenderObjectElement
+    with NotifiableElementMixin, ViewportElementMixin {
+  _PageViewCrossAxisShrinkWrappingViewportElement(
+    _PageViewCrossAxisShrinkWrappingViewport super.widget,
+  );
+
+  @override
+  void debugVisitOnstageChildren(ElementVisitor visitor) {
+    children
+        .where((Element child) {
+          final RenderSliver renderSliver = child.renderObject! as RenderSliver;
+          return renderSliver.geometry!.visible;
+        })
+        .forEach(visitor);
+  }
+}
+
+class _RenderPageViewCrossAxisShrinkWrappingViewport extends RenderShrinkWrappingViewport {
+  _RenderPageViewCrossAxisShrinkWrappingViewport({
+    super.axisDirection,
+    required super.crossAxisDirection,
+    required super.offset,
+    super.clipBehavior,
+    super.scrollCacheExtent,
+    List<RenderSliver>? children,
+  }) : super(children: children);
+
+  @override
+  bool debugThrowIfNotCheckingIntrinsics() {
+    assert(() {
+      if (!RenderObject.debugCheckingIntrinsics) {
+        throw FlutterError.fromParts(<DiagnosticsNode>[
+          ErrorSummary('$runtimeType does not support returning intrinsic dimensions.'),
+          ErrorDescription(
+            'Calculating the intrinsic dimensions would require instantiating every child of '
+            'the viewport, which defeats the point of viewports being lazy.',
+          ),
+          ErrorHint(
+            'If you are merely trying to shrink-wrap the viewport in the cross axis direction, '
+            'you should be able to achieve that effect by just giving the viewport loose '
+            'constraints, without needing to measure its intrinsic dimensions.',
+          ),
+        ]);
+      }
+      return true;
+    }());
+    return true;
+  }
+
+  late double _maxScrollExtent;
+  late double _shrinkWrapExtent;
+  late double _lastLaidOutCrossAxisExtent;
+  bool _hasReportedShrinkWrapExtent = false;
+  bool _hasVisualOverflow = false;
+
+  // Local cache extent field. The base class's _calculatedCacheExtent is
+  // library-private, so we maintain our own and override describeSemanticsClip.
+  double? _localCalculatedCacheExtent;
+
+  double _calculateCacheOffset(double mainAxisExtent) {
+    return switch (scrollCacheExtent.style) {
+      CacheExtentStyle.pixel => scrollCacheExtent.value,
+      CacheExtentStyle.viewport => scrollCacheExtent.value * mainAxisExtent,
+    };
+  }
+
+  bool _debugCheckHasBoundedMainAxis() {
+    assert(() {
+      switch (axis) {
+        case Axis.vertical:
+          if (!constraints.hasBoundedHeight) {
+            throw FlutterError(
+              'Vertical viewport was given unbounded height.\n'
+              'Cross-axis-shrinkwrapping viewports expand in the main axis to fill their '
+              'container and constrain their children to match their extent in the main axis. '
+              'In this case, a vertical cross-axis-shrinkwrapping viewport was given an '
+              'unlimited amount of vertical space in which to expand.',
+            );
+          }
+        case Axis.horizontal:
+          if (!constraints.hasBoundedWidth) {
+            throw FlutterError(
+              'Horizontal viewport was given unbounded width.\n'
+              'Cross-axis-shrinkwrapping viewports expand in the main axis to fill their '
+              'container and constrain their children to match their extent in the main axis. '
+              'In this case, a horizontal cross-axis-shrinkwrapping viewport was given an '
+              'unlimited amount of horizontal space in which to expand.',
+            );
+          }
+      }
+      return true;
+    }());
+    return true;
+  }
+
+  @override
+  Rect? describeSemanticsClip(RenderSliver? child) {
+    if (child != null &&
+        child.ensureSemantics &&
+        !(child.geometry!.visible || child.geometry!.cacheExtent > 0.0)) {
+      return null;
+    }
+    if (_localCalculatedCacheExtent == null) {
+      return semanticBounds;
+    }
+    return switch (axis) {
+      Axis.vertical => Rect.fromLTRB(
+        semanticBounds.left,
+        semanticBounds.top - _localCalculatedCacheExtent!,
+        semanticBounds.right,
+        semanticBounds.bottom + _localCalculatedCacheExtent!,
+      ),
+      Axis.horizontal => Rect.fromLTRB(
+        semanticBounds.left - _localCalculatedCacheExtent!,
+        semanticBounds.top,
+        semanticBounds.right + _localCalculatedCacheExtent!,
+        semanticBounds.bottom,
+      ),
+    };
+  }
+
+  @override
+  void performLayout() {
+    final BoxConstraints constraints = this.constraints;
+    assert(_debugCheckHasBoundedMainAxis());
+
+    final (double mainAxisExtent, double crossAxisExtent) = switch (axis) {
+      Axis.vertical => (constraints.maxHeight, constraints.maxWidth),
+      Axis.horizontal => (constraints.maxWidth, constraints.maxHeight),
+    };
+
+    offset.applyViewportDimension(mainAxisExtent);
+
+    if (firstChild == null) {
+      size = switch (axis) {
+        Axis.vertical => Size(constraints.minWidth, mainAxisExtent),
+        Axis.horizontal => Size(mainAxisExtent, constraints.minHeight),
+      };
+      _maxScrollExtent = 0.0;
+      _shrinkWrapExtent = 0.0;
+      _hasVisualOverflow = false;
+      offset.applyContentDimensions(0.0, 0.0);
+      return;
+    }
+
+    double correction;
+    while (true) {
+      correction = _attemptLayout(mainAxisExtent, crossAxisExtent, offset.pixels);
+      if (correction != 0.0) {
+        offset.correctBy(correction);
+      } else {
+        final bool didAcceptContentDimension = offset.applyContentDimensions(
+          0.0,
+          math.max(0.0, _maxScrollExtent - mainAxisExtent),
+        );
+        if (didAcceptContentDimension) {
+          break;
+        }
+      }
+    }
+
+    final double effectiveExtent = switch (axis) {
+      Axis.vertical => constraints.constrainWidth(
+        _hasReportedShrinkWrapExtent
+            ? _shrinkWrapExtent
+            : (constraints.maxWidth.isFinite ? constraints.maxWidth : 0.0),
+      ),
+      Axis.horizontal => constraints.constrainHeight(
+        _hasReportedShrinkWrapExtent
+            ? _shrinkWrapExtent
+            : (constraints.maxHeight.isFinite ? constraints.maxHeight : 0.0),
+      ),
+    };
+
+    size = switch (axis) {
+      Axis.vertical => constraints.constrainDimensions(effectiveExtent, mainAxisExtent),
+      Axis.horizontal => constraints.constrainDimensions(mainAxisExtent, effectiveExtent),
+    };
+  }
+
+  double _attemptLayout(double mainAxisExtent, double crossAxisExtent, double correctedOffset) {
+    assert(!mainAxisExtent.isNaN);
+    assert(mainAxisExtent >= 0.0);
+    assert(mainAxisExtent.isFinite);
+    assert(!crossAxisExtent.isNaN);
+    assert(crossAxisExtent >= 0.0);
+    assert(correctedOffset.isFinite);
+    _maxScrollExtent = 0.0;
+    _shrinkWrapExtent = 0.0;
+    _lastLaidOutCrossAxisExtent = crossAxisExtent;
+    _hasReportedShrinkWrapExtent = false;
+    _hasVisualOverflow = correctedOffset < 0.0;
+
+    _localCalculatedCacheExtent = _calculateCacheOffset(mainAxisExtent);
+
+    return layoutChildSequence(
+      child: firstChild,
+      scrollOffset: math.max(0.0, correctedOffset),
+      overlap: math.min(0.0, correctedOffset),
+      layoutOffset: math.max(0.0, -correctedOffset),
+      remainingPaintExtent: mainAxisExtent + math.min(0.0, correctedOffset),
+      mainAxisExtent: mainAxisExtent,
+      crossAxisExtent: crossAxisExtent,
+      growthDirection: GrowthDirection.forward,
+      advance: childAfter,
+      remainingCacheExtent: mainAxisExtent + 2 * _localCalculatedCacheExtent!,
+      cacheOrigin: -_localCalculatedCacheExtent!,
+    );
+  }
+
+  @override
+  bool get hasVisualOverflow => _hasVisualOverflow;
+
+  @override
+  void updateOutOfBandData(GrowthDirection growthDirection, SliverGeometry childLayoutGeometry) {
+    assert(growthDirection == GrowthDirection.forward);
+    _maxScrollExtent += childLayoutGeometry.scrollExtent;
+    if (childLayoutGeometry.hasVisualOverflow) {
+      _hasVisualOverflow = true;
+    }
+    if (!childLayoutGeometry.visible) {
+      return;
+    }
+    final double? childCrossAxisExtent = childLayoutGeometry.crossAxisExtent;
+    if (childCrossAxisExtent == null) {
+      if (!_lastLaidOutCrossAxisExtent.isFinite) {
+        return;
+      }
+      _hasReportedShrinkWrapExtent = true;
+      _shrinkWrapExtent = math.max(_shrinkWrapExtent, _lastLaidOutCrossAxisExtent);
+      return;
+    }
+    assert(childCrossAxisExtent.isFinite);
+    _hasReportedShrinkWrapExtent = true;
+    _shrinkWrapExtent = math.max(_shrinkWrapExtent, childCrossAxisExtent);
   }
 }
