+++ packages/flutter/lib/src/widgets/viewport.dart
@@ -526,189 +526,3 @@ class ShrinkWrappingViewport extends MultiChildRenderObjectWidget {
     );
   }
 }
-
-/// A widget through which a portion of larger content can be viewed, typically
-/// in combination with a [Scrollable], that shrink-wraps its contents in the
-/// cross axis.
-///
-/// [CrossAxisShrinkWrappingViewport] differs from [Viewport] in that
-/// [Viewport] expands to fill the cross axis whereas
-/// [CrossAxisShrinkWrappingViewport] sizes itself to match its children in the
-/// cross axis. This shrink wrapping behavior is expensive because the children,
-/// and hence the viewport, could potentially change size whenever the [offset]
-/// changes.
-///
-/// [CrossAxisShrinkWrappingViewport] cannot contain box children directly.
-/// Instead, use a [SliverList], [SliverFixedExtentList], [SliverGrid], or a
-/// [SliverToBoxAdapter], for example.
-///
-/// See also:
-///
-///  * [ListView], [PageView], [GridView], and [CustomScrollView], which combine
-///    [Scrollable] and a viewport into widgets that are easier to use.
-///  * [ShrinkWrappingViewport], a viewport that shrink-wraps in the main axis.
-///  * [Viewport], a viewport that does not shrink-wrap its contents.
-class CrossAxisShrinkWrappingViewport extends MultiChildRenderObjectWidget {
-  /// Creates a widget that is bigger on the inside and shrink-wraps its
-  /// children in the cross axis.
-  ///
-  /// The viewport listens to the [offset], which means you do not need to
-  /// rebuild this widget when the [offset] changes.
-  const CrossAxisShrinkWrappingViewport({
-    super.key,
-    this.axisDirection = AxisDirection.down,
-    this.crossAxisDirection,
-    required this.offset,
-    this.paintOrder = SliverPaintOrder.firstIsTop,
-    this.clipBehavior = Clip.hardEdge,
-    @Deprecated(
-      'Use scrollCacheExtent instead. '
-      'This feature was deprecated after v3.41.0-0.0.pre.',
-    )
-    this.cacheExtent,
-    @Deprecated(
-      'Use scrollCacheExtent instead. '
-      'This feature was deprecated after v3.41.0-0.0.pre.',
-    )
-    this.cacheExtentStyle = CacheExtentStyle.pixel,
-    this.scrollCacheExtent,
-    List<Widget> slivers = const <Widget>[],
-  }) : super(children: slivers);
-
-  /// The direction in which the [offset]'s [ViewportOffset.pixels] increases.
-  ///
-  /// For example, if the [axisDirection] is [AxisDirection.down], a scroll
-  /// offset of zero is at the top of the viewport and increases towards the
-  /// bottom of the viewport.
-  final AxisDirection axisDirection;
-
-  /// The direction in which child should be laid out in the cross axis.
-  ///
-  /// If the [axisDirection] is [AxisDirection.down] or [AxisDirection.up], this
-  /// property defaults to [AxisDirection.left] if the ambient [Directionality]
-  /// is [TextDirection.rtl] and [AxisDirection.right] if the ambient
-  /// [Directionality] is [TextDirection.ltr].
-  ///
-  /// If the [axisDirection] is [AxisDirection.left] or [AxisDirection.right],
-  /// this property defaults to [AxisDirection.down].
-  final AxisDirection? crossAxisDirection;
-
-  /// Which part of the content inside the viewport should be visible.
-  ///
-  /// The [ViewportOffset.pixels] value determines the scroll offset that the
-  /// viewport uses to select which part of its content to display. As the user
-  /// scrolls the viewport, this value changes, which changes the content that
-  /// is displayed.
-  ///
-  /// Typically a [ScrollPosition].
-  final ViewportOffset offset;
-
-  /// {@macro flutter.rendering.RenderViewportBase.paintOrder}
-  ///
-  /// Defaults to [SliverPaintOrder.firstIsTop].
-  final SliverPaintOrder paintOrder;
-
-  /// {@macro flutter.material.Material.clipBehavior}
-  ///
-  /// Defaults to [Clip.hardEdge].
-  final Clip clipBehavior;
-
-  /// {@macro flutter.rendering.RenderViewportBase.cacheExtent}
-  ///
-  /// See also:
-  ///
-  ///  * [cacheExtentStyle], which controls the units of the [cacheExtent].
-  @Deprecated(
-    'Use scrollCacheExtent instead. '
-    'This feature was deprecated after v3.41.0-0.0.pre.',
-  )
-  final double? cacheExtent;
-
-  /// {@macro flutter.rendering.RenderViewportBase.cacheExtentStyle}
-  @Deprecated(
-    'Use scrollCacheExtent instead. '
-    'This feature was deprecated after v3.41.0-0.0.pre.',
-  )
-  final CacheExtentStyle cacheExtentStyle;
-
-  /// {@macro flutter.rendering.RenderViewportBase.scrollCacheExtent}
-  final ScrollCacheExtent? scrollCacheExtent;
-
-  ScrollCacheExtent? get _effectiveScrollCacheExtent {
-    if (scrollCacheExtent != null) {
-      return scrollCacheExtent;
-    }
-    if (cacheExtent != null) {
-      switch (cacheExtentStyle) {
-        case CacheExtentStyle.pixel:
-          return ScrollCacheExtent.pixels(cacheExtent!);
-        case CacheExtentStyle.viewport:
-          return ScrollCacheExtent.viewport(cacheExtent!);
-      }
-    }
-    return null;
-  }
-
-  @override
-  MultiChildRenderObjectElement createElement() => _CrossAxisShrinkWrappingViewportElement(this);
-
-  @override
-  RenderCrossAxisShrinkWrappingViewport createRenderObject(BuildContext context) {
-    return RenderCrossAxisShrinkWrappingViewport(
-      axisDirection: axisDirection,
-      crossAxisDirection:
-          crossAxisDirection ?? Viewport.getDefaultCrossAxisDirection(context, axisDirection),
-      offset: offset,
-      paintOrder: paintOrder,
-      clipBehavior: clipBehavior,
-      scrollCacheExtent: _effectiveScrollCacheExtent,
-    );
-  }
-
-  @override
-  void updateRenderObject(
-    BuildContext context,
-    RenderCrossAxisShrinkWrappingViewport renderObject,
-  ) {
-    renderObject
-      ..axisDirection = axisDirection
-      ..crossAxisDirection =
-          crossAxisDirection ?? Viewport.getDefaultCrossAxisDirection(context, axisDirection)
-      ..offset = offset
-      ..paintOrder = paintOrder
-      ..clipBehavior = clipBehavior
-      ..scrollCacheExtent = _effectiveScrollCacheExtent;
-  }
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties.add(EnumProperty<AxisDirection>('axisDirection', axisDirection));
-    properties.add(
-      EnumProperty<AxisDirection>('crossAxisDirection', crossAxisDirection, defaultValue: null),
-    );
-    properties.add(DiagnosticsProperty<ViewportOffset>('offset', offset));
-    properties.add(
-      DiagnosticsProperty<ScrollCacheExtent>(
-        'scrollCacheExtent',
-        scrollCacheExtent,
-        defaultValue: null,
-      ),
-    );
-  }
-}
-
-class _CrossAxisShrinkWrappingViewportElement extends MultiChildRenderObjectElement
-    with NotifiableElementMixin, ViewportElementMixin {
-  _CrossAxisShrinkWrappingViewportElement(CrossAxisShrinkWrappingViewport super.widget);
-
-  @override
-  void debugVisitOnstageChildren(ElementVisitor visitor) {
-    children
-        .where((Element child) {
-          final RenderSliver renderSliver = child.renderObject! as RenderSliver;
-          return renderSliver.geometry!.visible;
-        })
-        .forEach(visitor);
-  }
-}
